cdk/requirements.txt
-e .
aws_cdk.aws_ec2
aws_cdk.aws_eks
aws_cdk.aws_iam
requests

